Add files via upload
The Number Guessing Game is a simple, interactive game where the player attempts to guess a randomly generated number within a limited number of chances. The game provides feedback on whether the player's guess is too high, too low, or correct. This game demonstrates basic concepts of programming such as loops, conditionals, and user input handling.
Some features : It randomly generates a number between 1 to 100 , taking the input number and match with the randomly generated number if the the number is small it too low , if number is greater then it shows too high , if number is not in the range it shows invalid number if both the number are equal then it shows you have guessed the correct number .And it allows the players to multiple rounds .